Moved CategoriesHolder from search into indexer library. Need to access it from Editor.
