+++ indexer/categories_holder.hpp
