+++ indexer/indexer_tests/indexer_tests.pro
@@ -17,6 +17,7 @@ HEADERS += \
 
 SOURCES += \
     ../../testing/testingmain.cpp \
+    categories_test.cpp \
     cell_coverer_test.cpp \
     cell_id_test.cpp \
     checker_test.cpp \
@@ -33,6 +34,7 @@ SOURCES += \
     point_to_int64_test.cpp \
     rank_table_test.cpp \
     scales_test.cpp \
+    search_string_utils_test.cpp \
     sort_and_merge_intervals_test.cpp \
     succinct_trie_test.cpp \
     test_polylines.cpp \
