+++ indexer/search_delimiters.cpp
