+++ indexer/search_string_utils.cpp
