+++ map/address_finder.cpp
@@ -1,11 +1,11 @@
 #include "map/framework.hpp"
 
-#include "search/categories_holder.hpp"
 #include "search/result.hpp"
 #include "search/reverse_geocoder.hpp"
 
 #include "drape_frontend/visual_params.hpp"
 
+#include "indexer/categories_holder.hpp"
 #include "indexer/classificator.hpp"
 #include "indexer/feature_algo.hpp"
 #include "indexer/feature_visibility.hpp"
