+++ search/approximate_string_match.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "search_string_utils.hpp"
+#include "indexer/search_string_utils.hpp"
 
 #include "base/base.hpp"
 #include "base/buffer_vector.hpp"
