+++ search/house_detector.cpp
@@ -2,10 +2,10 @@
 
 #include "search/algos.hpp"
 #include "search/search_common.hpp"
-#include "search/search_string_utils.hpp"
 
 #include "indexer/classificator.hpp"
 #include "indexer/feature_impl.hpp"
+#include "indexer/search_string_utils.hpp"
 
 #include "geometry/angles.hpp"
 #include "geometry/distance.hpp"
