+++ search/keyword_matcher.cpp
@@ -1,7 +1,7 @@
 #include "keyword_matcher.hpp"
 
-#include "search_delimiters.hpp"
-#include "search_string_utils.hpp"
+#include "indexer/search_delimiters.hpp"
+#include "indexer/search_string_utils.hpp"
 
 #include "base/stl_add.hpp"
