+++ search/search.pro
@@ -12,7 +12,6 @@ HEADERS += \
     algos.hpp \
     approximate_string_match.hpp \
     cancel_exception.hpp \
-    categories_holder.hpp \
     dummy_rank_table.hpp \
     feature_offset_match.hpp \
     geometry_utils.hpp \
@@ -34,14 +33,12 @@ HEADERS += \
     retrieval.hpp \
     reverse_geocoder.hpp \
     search_common.hpp \
-    search_delimiters.hpp \
     search_engine.hpp \
     search_index_values.hpp \
     search_query.hpp \
     search_query_factory.hpp \
     search_query_params.hpp \
     search_string_intersection.hpp \
-    search_string_utils.hpp \
     search_trie.hpp \
     suggest.hpp \
     v2/cbv_ptr.hpp \
@@ -63,7 +60,6 @@ HEADERS += \
 
 SOURCES += \
     approximate_string_match.cpp \
-    categories_holder.cpp \
     dummy_rank_table.cpp \
     geometry_utils.cpp \
     house_detector.cpp \
@@ -81,11 +77,9 @@ SOURCES += \
     result.cpp \
     retrieval.cpp \
     reverse_geocoder.cpp \
-    search_delimiters.cpp \
     search_engine.cpp \
     search_query.cpp \
     search_query_params.cpp \
-    search_string_utils.cpp \
     v2/cbv_ptr.cpp \
     v2/features_filter.cpp \
     v2/features_layer.cpp \
