+++ search/search_engine.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
-#include "categories_holder.hpp"
 #include "params.hpp"
 #include "result.hpp"
 #include "search_query_factory.hpp"
 #include "suggest.hpp"
+
+#include "indexer/categories_holder.hpp"
 
 #include "geometry/rect2d.hpp"
