+++ search/search_query.cpp
@@ -1,6 +1,5 @@
 #include "search_query.hpp"
 
-#include "search/categories_holder.hpp"
 #include "search/dummy_rank_table.hpp"
 #include "search/feature_offset_match.hpp"
 #include "search/geometry_utils.hpp"
@@ -9,14 +8,13 @@
 #include "search/locality.hpp"
 #include "search/region.hpp"
 #include "search/search_common.hpp"
-#include "search/search_delimiters.hpp"
 #include "search/search_index_values.hpp"
 #include "search/search_query_params.hpp"
 #include "search/search_string_intersection.hpp"
-#include "search/search_string_utils.hpp"
 
 #include "storage/country_info_getter.hpp"
 
+#include "indexer/categories_holder.hpp"
 #include "indexer/classificator.hpp"
 #include "indexer/feature.hpp"
 #include "indexer/feature_algo.hpp"
@@ -25,6 +23,8 @@
 #include "indexer/features_vector.hpp"
 #include "indexer/index.hpp"
 #include "indexer/scales.hpp"
+#include "indexer/search_delimiters.hpp"
+#include "indexer/search_string_utils.hpp"
 #include "indexer/trie_reader.hpp"
 
 #include "platform/mwm_traits.hpp"
