+++ search/search_string_intersection.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "search_delimiters.hpp"
+#include "indexer/search_delimiters.hpp"
 
 
 namespace search
