+++ search/search_tests/house_detector_tests.cpp
@@ -1,18 +1,17 @@
 #include "testing/testing.hpp"
 
 #include "search/house_detector.hpp"
-#include "search/search_string_utils.hpp"
 
 #include "indexer/classificator_loader.hpp"
 #include "indexer/data_header.hpp"
 #include "indexer/ftypes_matcher.hpp"
 #include "indexer/index.hpp"
 #include "indexer/scales.hpp"
-
-#include "platform/platform.hpp"
+#include "indexer/search_string_utils.hpp"
 
 #include "platform/local_country_file.hpp"
 #include "platform/local_country_file_utils.hpp"
+#include "platform/platform.hpp"
 
 #include "geometry/distance_on_sphere.hpp"
