+++ search/search_tests/keyword_matcher_test.cpp
@@ -2,8 +2,9 @@
 
 #include "search/keyword_matcher.hpp"
 #include "search/search_common.hpp"
-#include "search/search_delimiters.hpp"
-#include "search/search_string_utils.hpp"
+
+#include "indexer/search_delimiters.hpp"
+#include "indexer/search_string_utils.hpp"
 
 #include "base/buffer_vector.hpp"
 #include "base/stl_add.hpp"
