+++ search/search_tests/string_match_test.cpp
@@ -1,7 +1,8 @@
 #include "testing/testing.hpp"
 
 #include "search/approximate_string_match.hpp"
-#include "search/search_delimiters.hpp"
+
+#include "indexer/search_delimiters.hpp"
 
 #include "search/search_tests/match_cost_mock.hpp"
